Centralize all arguments parsing in lib.rs
